Split general manager tests from Docker credential manager tests
