+++ laituri_tests/mock_data.py
@@ -1,0 +1,8 @@
+EXAMPLE_IMAGES = (
+    'docker.io/owner/project:tag',
+    'docker.io/owner/project',
+    'owner/project:tag',
+    'owner/project',
+    'example.com/owner/project:tag',
+    'example.com/owner/project',
+)
